Assign known people to the support project
When a Productive person is not assigned to the support project the API
call that books them will fail (with an unhelpful 403 error), to work
around this we check if the person being booked is assigned to the
support project and assign them if not before we attempt to make the
booking.

The API call that looks for assignments seems to make two sub calls so
we have to stub those out as well.

Co-authored-by: Laura <laura@dxw.com>