Rename Booking to SendToProductive
Because Productive has a Booking object, and we have a Booking object, the code
was getting slightly confusing as we were finding it difficult to distinguish
between "our" booking and "their" booking.

This change originated in breathe_to_productive and we are replicating it here.
"Our" Booking is now called SendToProductive to imply it is a data object that
is being sent into Productive to be turned into a Booking.