Decide which rotas to import via environment variables
At the moment we're importing the in-hours ops and developer support rotas by
default. We might in future decide not to add both, or to add more rotas
(e.g the out of hours rotas). In order to make this easier, add environment
variables to indicate which rotas we want to import.