Test suite can use different Support Rota fixtures
The previous fixtures were fixed so that there was one dev and one ops. These fixtures included a total of 4 events. Every test that uses fixtures then had to set up a world where there were 4 employees, with person records etc. This made understanding the tests difficult and encouraged the use of lets to hide the problem.

Instead we will allow different fixtures to be used. We can have empty fixtures to reduce test set up and help clarify what is the data that's subject of the test for easier readability.