Prevent duplicate Productive support bookings by a filtering out 2 months of past support rota data
Before this change we looked at the oldest event returned by the support rota to decide how many Productive support bookings we wanted to query for and compare againt.

2 months is a long time. Though it has been useful to create older records on the initial run we don't think we need to keep this around. We think that corrections will only be needed last minute if a person on support is sick or needs to be covered unexpectedly. We think that this correction is likely to be made in OpsGenie on the same day. In which case, the next time this task is run (which is 3am), we need to check the day before again to make sure there were no corrections.

This change reduces the coupling between support_rota responses and productive queries which simplifies the testing approach.

We stop using Factories to stub our test suite. This worked in the background to stub a record everytime we created it. Whilst convinent in our feature tests this has 3 problems:

1. It's hidden stubbing that the user isn't aware of, this makes it hard to change
2. We don't always want to stub a booking when we create one, our test suite can be full of stubs that are never used, this makes it hard to work on since it's hard to know what's actually used
3. It's hard to extend as is the case. We want to change a query to change the filter date that's used in this stub. It was hardcoded to today and had no way to inject an attribute from another record into it.

We move the JSON building part of what the factory was doing into the stub methods.