Avoid creating and deleting all records on each run
Before this change the comparison of what was in productive vs what was in the support rota wasn't working as expected: `common_support_rotations()` always returned `[]`.

This is because the object comparison being done between objects found in each API were never matching.

The support rota `date` field was a String and the productive `date` field was a `date`.

This change makes the support rota "date", a Date, rather than a String.