Stubbing of People is done with explicit test helpers
We needed to test the scenario where Support rota email addresses **did not** match the email address in Productive.

The previous FactoryBot use of transient, for convience blocks this as it's automatically plumbed in to stub people and employees with the same values.

The test suite is also refactored to use less `let` to help make the code easier to reasion about before making this fix.

We stop building 5 employees in the "this person doesn't exist in Productive" context since they were never used.

We weave `stub_people` into more tests to explicitly state the expected response from Productive.