Stop deleting and recreating all support events with different email casing
Before this change the script would incorrectly identify that there were x bookings in Productive that needed deleting.

Users in the Support rota all have fully lowercased email addresses.

When creating bookings we use the productive email address which includes this uppercase.

When we run the script again, we compare email addresses from the support rota and as they're lowercase, it detects a difference. It deletes them all, then recreates them all.